refactor

+++ bench/change_model.py
@@ -1004,7 +1004,7 @@ def estimate_median(log_lh: Callable, bounds, n_samples=int(1e3)):
     :return:
     """
     x = np.linspace(bounds[0], bounds[1], n_samples)
-    lh = np.array([0] + [np.exp(log_lh(x_)) for x_ in x])
+    lh = np.array([0.] + [np.exp(log_lh(x_)) for x_ in x], dtype=float)
     p_idx = np.argwhere(np.cumsum(lh) / lh.sum() < 0.5)[-1]
     return x[p_idx]
